refactors redis cache check and insertion
