+++ tesseract-server/src/handlers/aggregate.rs
@@ -20,7 +20,11 @@ use crate::handlers::util::validate_members;
 
 use crate::app::AppState;
 use crate::errors::ServerError;
-use super::util::{boxed_error_http_response, verify_authorization, format_to_content_type, generate_source_data};
+use super::util::{
+    boxed_error_http_response, verify_authorization,
+    format_to_content_type, generate_source_data,
+    get_redis_cache_key, check_redis_cache, insert_into_redis_cache
+};
 use r2d2_redis::{redis};
 
 /// Handles default aggregation when a format is not specified.
@@ -73,9 +77,16 @@ pub fn do_aggregate(
 
     info!("query opts:{:?}", agg_query);
 
+    // Check if this query is already cached
+    let redis_pool = req.state().redis_pool.clone();
+    let redis_cache_key = get_redis_cache_key(&req, &format);
+
+    if let Some(res) = check_redis_cache(&format, &redis_pool, &redis_cache_key) {
+        return res;
+    }
+
     // Gets the Source Data
     let source_data = Some(generate_source_data(&cube_obj));
-
 
     // Turn AggregateQueryOpt into Query
     let ts_query: Result<TsQuery, _> = agg_query.try_into();
@@ -101,28 +112,6 @@ pub fn do_aggregate(
     info!("Sql query: {}", sql);
     info!("Headers: {:?}", headers);
     
-    let redis_pool = req.state().redis_pool.clone();
-    let cache_key = format!("{}-{:?}", &sql, &format);
-    // Check if SQL is in Redis
-    if let Some(rpool) =  &redis_pool {
-        let conn_result = rpool.get();
-        // Check cache...
-        if let Ok(mut conn) = conn_result {
-            let redis_cache_result = redis::cmd("GET").arg(&cache_key).query(&mut *conn);
-            if let Ok(result_str) = redis_cache_result {
-                let result_str: &String = &result_str;
-                let content_type = format_to_content_type(&format);
-                let response = HttpResponse::Ok()
-                .set(content_type)
-                    .body(result_str);
-                return Box::new(future::result(Ok(response)));
-            }
-        } else {
-            debug!("Failed to get redis pool handle!");
-        }
-        // cache miss!
-    }
-
     req.state()
         .backend
         .exec_sql(sql)
@@ -131,14 +120,9 @@ pub fn do_aggregate(
 
             match format_records(&headers, df, format, source_data, false) {
                 Ok(res) => {
-                    if let Some(rpool) = &redis_pool { // if user set redis 
-                        if let Ok(mut conn) = rpool.get() { // if able to get pool handle
-                            let rs: redis::RedisResult<String> = redis::cmd("SET").arg(&cache_key).arg(&res).query(&mut *conn);
-                            if rs.is_err() {
-                                debug!("Error occurred when trying to save key: {}", &cache_key);
-                            }
-                        }
-                    }
+                    // Try to insert this result in the Redis cache, if available
+                    insert_into_redis_cache(&res, &redis_pool, &redis_cache_key);
+
                     Ok(HttpResponse::Ok()
                         .set(content_type)
                         .body(res))
